requirements.txt
streamlit==1.33.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pygments==2.19.2
